Add membershipstatus
The members lacked a status field which made it hard
to keep track of whether members were ex-members, members,
expelled, cancelled their membership etc.

This introduces arbitrary fields of membership statuses
that are being able to be set by the administrators.

The migration makes a single membership type in advance
which is called "Lid" (member) and everyone by default
will be made member and get access to the login.

You can make additional membership types which can't
login to the system at all.