Add membershipStatus to member revision (#157)
This way history is kept for changing memberships, which is
interesting to view membership history over time.

The event listener responsible for this was already implemented
in: src/EventSubscriber/EasyAdminEventSubscriber.php as it
simply called the constructor of the MemberDetailsRevision
and passes a Membership as its argument. Therefore just
adding the field, a migration and changing the constructor
to add this makes this work.