[144] Make contributions dynamic
Also fixed the validation which checks that the higher amount is higher
than the highest tier. And fixed the clearing/setting of the higher
amount radio button, which was working on the signup page but not on the
contribution page, so I just copied it from the signup page and it
worked.