[239] Move general Group fields back to top
Apparently fields that are not part of a form panel are moved to the
other fields which are part of panels, so the most important fields were
moved to the bottom. This in turn made it very hard to set the contacts
of the group due to the select list being at the bottom of the screen.